Bump all dependencies to latest versions
- Spring Boot parent 4.0.3 -> 4.1.1 (caffeine 3.2.4, flyway 12.4.0,
  postgresql 42.7.13 managed versions ride along)
- compiler-plugin 3.16.0, lombok 1.18.48 (via lombok.version property),
  httpclient5 5.6.4, httpcore5 5.4.3, okhttp 5.5.0, github-api 2.0-rc.7,
  adventure 5.2.0, guava 33.7.1-jre, springdoc 3.1.1, geoip2 5.2.0,
  plexus-archiver 4.14.0, minio 9.0.3, prometheus-metrics 1.9.0
- dnsjava: jitpack com.github.dnsjava:v3.5.2 -> central dnsjava:3.6.5,
  remove jitpack.io repository
- minio 9: InetAddressValidator import moved to io.minio.Utils;
  PutObjectArgs.stream sizes now boxed Long

+++ src/main/java/xyz/mcutils/backend/service/StorageService.java
@@ -55,7 +55,7 @@ public class StorageService {
     public void upload(Bucket bucket, String fileName, String contentType, byte[] data) {
         long before = System.currentTimeMillis();
         try {
-            this.minioClient.putObject(PutObjectArgs.builder().bucket(bucket.getName()).object(fileName).stream(new ByteArrayInputStream(data), data.length, -1).contentType(contentType)
+            this.minioClient.putObject(PutObjectArgs.builder().bucket(bucket.getName()).object(fileName).stream(new ByteArrayInputStream(data), (long) data.length, -1L).contentType(contentType)
 
                     .build());
             if (this.objectCache != null) {
